backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==23.0.0
supabase==2.15.3
python-dotenv==1.1.1
redis==5.2.1
